TruckTest.java
import static org.junit.Assert.*;
import org.junit.Test;

public class TruckTest {

    // Test case for constructor
    @Test
    public void testConstructor() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_001", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        assertNotNull(truck);
    }

    // Test case for getLateFee method with positive difference
    @Test
    public void testGetLateFee_PositiveDifference() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_001", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        DateTime endDate = new DateTime(2000, 2, 12);
        DateTime startDate = new DateTime(2000, 2, 10);
        assertEquals(187.5, truck.getLateFee(endDate, startDate), 0.001);
    }

    // Test case for getLateFee method with zero difference
    @Test
    public void testGetLateFee_ZeroDifference() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_002", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        DateTime endDate = new DateTime(2000, 2, 12);
        DateTime startDate = new DateTime(2000, 2, 12);
        assertEquals(0.0, truck.getLateFee(endDate, startDate), 0.001);
    }

    // Test case for returnVehicle method when vehicle is available
    @Test
    public void testReturnVehicle_Available() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_003", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        assertFalse(truck.returnVehicle(new DateTime(2000, 2, 12)));
    }

    // Test case for completeMaintenance method when vehicle is not in maintenance
    @Test
    public void testCompleteMaintenance_NotInMaintenance() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_004", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        assertFalse(truck.completeMaintenance());
    }

    // Test case for completeMaintenance method when vehicle is in maintenance
    @Test
    public void testCompleteMaintenance_InMaintenance() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_005", 2022, "Ford", "F-150", 2, truckType, "Blue", "John Doe");
        assertTrue(truck.completeMaintenance());
    }

    // Test case for getDetails method with empty rental record
    @Test
    public void testGetDetails_EmptyRentalRecord() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_006", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        assertEquals("\nRENTAL RECORD:\tempty", truck.getDetails());
    }

    // Test case for getDetails method with non-empty rental record
    @Test
    public void testGetDetails_NonEmptyRentalRecord() {
        VehicleType truckType = new VehicleType();
        Truck truck = new Truck("Truck_007", 2022, "Ford", "F-150", 0, truckType, "Blue", "John Doe");
        RentalRecord record = new RentalRecord();
        String expectedDetails = "Details of rental record\n" +
                                 "--------------------------\n" +
                                 "Rental details                     \n";
        assertEquals(expectedDetails, truck.getDetails());
    }


	@Test
	public void testVehicleStringIntStringStringIntVehicleTypeStringString() {
		fail("Not yet implemented");
	}

	@Test
	public void testVehicleStringIntStringStringIntVehicleType() {
		fail("Not yet implemented");
	}

	@Test
	public void testGetVehicleId() {
		fail("Not yet implemented");
	}

	@Test
	public void testRent() {
		fail("Not yet implemented");
	}

	@Test
	public void testPerformMaintenance() {
		fail("Not yet implemented");
	}

	@Test
	public void testToString() {
		fail("Not yet implemented");
	}

	@Test
	public void testGetLastElementIndex() {
		fail("Not yet implemented");
	}

}
